% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{abre_arquivo}
\alias{abre_arquivo}
\title{Abre um arquivo usando o programa padrão do sistema operacional}
\usage{
abre_arquivo(arquivo)
}
\arguments{
\item{arquivo}{Uma string com o caminho do arquivo.}
}
\description{
Abre arquivos no Linux e no Windows
}
\details{
Se não funcionar, use o caminho completo (em vez do caminho relativo).
}
