% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{abre_tabela_excel}
\alias{abre_tabela_excel}
\title{Abre uma tabela no Excel}
\usage{
abre_tabela_excel(x, nome_opcional = NULL)
}
\arguments{
\item{x}{Dataframe}

\item{nome_opcional}{Nome opctional do arquivo. Se nulo, vai como tabela.xlsx}
}
\description{
Abre arquivos no Linux e no Windows
}
