% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{anonymize}
\alias{anonymize}
\title{Anonymize}
\usage{
anonymize(tables, columns, seed = NULL)
}
\arguments{
\item{tables}{Dataframe list}

\item{columns}{Columns to anonymize}
}
\value{
Original list, but with anonymized columns. If there were columns of same name in different tables, elements in common will have the same anonymization
}
\description{
Given a list of tables, anonimyzes the specified columns
}
\examples{
anonymize(list(iris), "Species")
}
