% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{arruma_eventos_da_tratativa}
\alias{arruma_eventos_da_tratativa}
\title{Escreve direito os eventos de uma tratativa}
\usage{
arruma_eventos_da_tratativa(tabela)
}
\arguments{
\item{tabela}{Tabela com os eventos esculachados.}
}
\value{
data.frame chiquíssimo.
}
\description{
Escreve direito os eventos de uma tratativa
}
