% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{atualiza_pacote}
\alias{atualiza_pacote}
\title{Auto atualizador do futuro}
\usage{
atualiza_pacote()
}
\value{
Retorna nada, só pega a versão mais recente do github
}
\description{
Auto atualizador do futuro
}
