% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{carrega_pacote}
\alias{carrega_pacote}
\title{Auto atualizador do futuro}
\usage{
carrega_pacote()
}
\value{
Retorna nada, só pega a versão mais recente do github e carrega.
}
\description{
Auto atualizador do futuro
}
