% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_etc.R
\name{collect_with_index}
\alias{collect_with_index}
\title{Dá collect numa query usando índices}
\usage{
collect_with_index(
  x,
  table_names = NULL,
  index_names = NULL,
  print_query = FALSE
)
}
\arguments{
\item{x}{Uma query usando o dbplyr}

\item{table_names}{Um vetor com o nome das tabelas usadas na query}

\item{index_names}{Um vetor com o nome dos índices a serem usados em cada tabela do vetor anterior}

\item{print_query}{Se TRUE, printa a query com o índice}
}
\value{
O collect da query
}
\description{
Dá collect numa query usando índices
}
