% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipular_strings.R
\name{compara_string}
\alias{compara_string}
\title{Checa se duas strings são iguais (sem dar pau com NA)}
\usage{
compara_string(x, y)
}
\arguments{
\item{x}{String a ser testada.}

\item{y}{String a ser testada, de preferência do mesmo tamanho que x.}
}
\value{
Vetor booleano do memo tamanho que  x
}
\description{
Checa se duas strings são iguais (sem dar pau com NA)
}
