% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{conecta_base_sqlite}
\alias{conecta_base_sqlite}
\title{Cria conexão com uma base do SQLite.}
\usage{
conecta_base_sqlite(local_database, nome_tabela)
}
\arguments{
\item{local_database}{Uma string com o caminho do database.}

\item{nome_tabela}{Nome da tabela do SQLite para acessar.}
}
\value{
Uma conexão com o SQLite.
}
\description{
Cria uma tabela 'simbólica' do SQLite.
}
\details{
Essa função retorna uma tabela
  na memória. Você poderá manipular ela como se estivesse de fato,
  usando os verbos do dplyr. Quando quiser realmente obter os dados
  após seus comandos, utilize 'collect()'.
}
