% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{consolida_base_tratativa_mariadb_nova}
\alias{consolida_base_tratativa_mariadb_nova}
\title{Pega eventos da base de tratativas do mariadb e sobe pra base de Analytics correspondente}
\usage{
consolida_base_tratativa_mariadb_nova(
  conexao_db_original,
  conexao_db_tratativas,
  db_selecionado,
  data_inicial = NULL,
  data_final = NULL
)
}
\arguments{
\item{conexao_db_original}{Conexão com o db original de Analytics}

\item{conexao_db_tratativas}{Conexão com o db de tratativas}

\item{db_selecionado}{Nome do db selecionado (pra saber quais operações da tratativa
ele tem que pegar)}

\item{data_inicial}{Se NULL, usa '2018-01-01'}

\item{data_final}{Se NULL, usa o dia de ontem}
}
\value{
Retorna nada fi
}
\description{
Pega eventos da base de tratativas do mariadb e sobe pra base de Analytics correspondente
}
