% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipular_strings.R
\name{converte_data_para_string}
\alias{converte_data_para_string}
\title{Data uma data, transforma em uma string no formato brasileiro. Tem opção de formatar pra mês/ano também.}
\usage{
converte_data_para_string(x, tipo = "Semana")
}
\arguments{
\item{x}{Vetor de datas.}

\item{tipo}{Semana ou mês. Se semana, fica dia/mês/ano. Se mês, fica mês/ano.}
}
\value{
Vetorzão da hora de characters.
}
\description{
Data uma data, transforma em uma string no formato brasileiro. Tem opção de formatar pra mês/ano também.
}
