% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dia_recente_da_semana}
\alias{dia_recente_da_semana}
\title{Procura a data mais recente em que ocorreu um dia da semana.}
\usage{
dia_recente_da_semana(
  data_a_considerar = lubridate::today(tz = "Brazil/East"),
  dia_da_semana = 0
)
}
\arguments{
\item{data_a_considerar}{Data a partir da qual procurar uma semana pra trás.}

\item{dia_da_semana}{Abreviado. Valores: seg, ter, qua, qui, sex, sáb, dom.}
}
\value{
Um objeto date com a data mais recente que cai no dia da semana selecionado.
}
\description{
Procura a data mais recente em que ocorreu um dia da semana.
}
