% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eh_erro}
\alias{eh_erro}
\title{Checa se o try é um erro}
\usage{
eh_erro(x)
}
\arguments{
\item{x}{Objeto a ser testado}
}
\value{
TRUE se for erro, FALSE se não for.
}
\description{
Checa se o try é um erro
}
