% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_etc.R
\name{escreve_mariadb}
\alias{escreve_mariadb}
\title{Escreve (dá append) numa base MariaDB}
\usage{
escreve_mariadb(dados_a_serem_salvos, nome_tabela, conexao, comando = "IGNORE")
}
\arguments{
\item{dados_a_serem_salvos}{A tabela com os dados a serem salvos no DB.}

\item{nome_tabela}{Nome da tabela no database onde os dados serão salvos.}

\item{conexao}{A conexão com o database.}

\item{comando}{Comando para o INSERT, as opções são IGNORE ou REPLACE.}
}
\description{
Escreve (dá append) numa base MariaDB
}
