% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{escreve_numa_base_mariadb2}
\alias{escreve_numa_base_mariadb2}
\title{Escreve (dá append) numa base MariaDB}
\usage{
escreve_numa_base_mariadb2(
  conexao,
  nome_tabela,
  dados_a_serem_salvos,
  comando = "IGNORE"
)
}
\arguments{
\item{conexao}{A conexão com o database.}

\item{nome_tabela}{Nome da tabela no database onde os dados serão salvos.}

\item{dados_a_serem_salvos}{A tabela com os dados a serem salvos no DB.}

\item{comando}{Comando para o INSERT, as opções são IGNORE ou REPLACE.}
}
\description{
Escreve (dá append) numa base MariaDB
}
