% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{escreve_numa_base_sqlite}
\alias{escreve_numa_base_sqlite}
\title{Escreve dados no SQLite.}
\usage{
escreve_numa_base_sqlite(
  dados_a_serem_salvos,
  local_database,
  nome_tabela,
  append = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{dados_a_serem_salvos}{A tabela a ser salva no SQLite.}

\item{local_database}{Uma string com o caminho do database.}

\item{nome_tabela}{Nome da tabela do SQLite para acessar.}

\item{append}{Se os dados serão adicionados ao fim da tabela do SQLite (default: TRUE)}

\item{overwrite}{Se os dados sobrescreverão o que já existe no SQLite (default: FALSE)}
}
\description{
Cria ou adiciona dados a uma tabela do SQLite
}
\details{
Use append = TRUE e overwrite = FALSE (padrão) para adicionar dados
a uma tabela já existente. Para sobrescrevê-la por completo, use
append = FALSE e overwrite = TRUE.
}
