% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{escreve_tabela_sql_geral}
\alias{escreve_tabela_sql_geral}
\title{Insere uma tabela no db via uma query gigante}
\usage{
escreve_tabela_sql_geral(.data, con, nome_tabela_db, comando = "REPLACE")
}
\arguments{
\item{.data}{Dataframe para ser escrito no banco}

\item{con}{Conexão com o database}

\item{nome_tabela_db}{Nome da tabela no database}

\item{comando}{Comando que vai ser usado na inserção}
}
\value{
O tanto de linhas que foi alterado (conta o insert e o update)
}
\description{
Insere uma tabela no db via uma query gigante
}
