% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_etc.R
\name{formata_string_sql}
\alias{formata_string_sql}
\title{Transforma um vetor de string em um texto formatado pro sql}
\usage{
formata_string_sql(x)
}
\arguments{
\item{x}{Vetor de strings}

\item{data_para_mexer}{Data a partir da qual fazer a resumida}
}
\value{
Um vetor de tamanho 1
}
\description{
Transforma um vetor de string em um texto formatado pro sql
}
