% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gera_calendario}
\alias{gera_calendario}
\title{Gera um calendário}
\usage{
gera_calendario(
  data_inicial = "2018-01-01",
  data_final = today(),
  dia_em_que_comeca_a_semana = "dom",
  semana = TRUE,
  mes = TRUE,
  ano = TRUE,
  dia_character = FALSE
)
}
\arguments{
\item{data_inicial}{Data inicial do calendário.}

\item{data_final}{Data final do calendário.}

\item{dia_em_que_comeca_a_semana}{Iniciais do dia da semana em que
a semana começa no calendário. Valores possívels:
'seg', 'ter', 'qua', 'qui', 'sex', 'sáb', 'dom'.}

\item{semana}{Se TRUE, o calendário terá coluna 'Semana'.}

\item{mes}{Se TRUE, o calendário terá coluna 'Mês'.}

\item{ano}{Se TRUE, o calendário terá coluna 'Ano'.}

\item{dia_character}{Se TRUE, a coluna 'Dia' vem como character.}
}
\value{
Um dataframe.
}
\description{
Gera uma tabela com dias, semanas, mês e ano, para usar
em algum left_join e assim agrupar seus dados por semana/mês/ano.
}
