% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gera_num_eventos_frota}
\alias{gera_num_eventos_frota}
\title{Gera o num_eventos_frota do período correspondente em diante}
\usage{
gera_num_eventos_frota(con, data_para_mexer = "2018-01-01")
}
\arguments{
\item{con}{Conexão com o database}

\item{data_para_mexer}{Data a partir da qual fazer a resumida}
}
\value{
Nadinha
}
\description{
Gera o num_eventos_frota do período correspondente em diante
}
