% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{instala_rdcomclient}
\alias{instala_rdcomclient}
\title{Instala o pacote RDCOMClient (que se comunica com o pacote Office)}
\usage{
instala_rdcomclient()
}
\description{
Instala o pacote RDCOMClient (que se comunica com o pacote Office)
}
