% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabelas_antigas.R
\name{junta_tratativas_numa_tabela_so}
\alias{junta_tratativas_numa_tabela_so}
\title{Pega a lista de tratativas e achata numa tabela só}
\usage{
junta_tratativas_numa_tabela_so(lista_de_tratativas)
}
\arguments{
\item{lista_de_tratativas}{Lista em que cada elemento é uma tabela.}
}
\value{
data.frame chique.
}
\description{
Pega a lista de tratativas e achata numa tabela só
}
