% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{le_tratativa_base_antiga}
\alias{le_tratativa_base_antiga}
\title{Lê a tratativa já com os nomes uniformes}
\usage{
le_tratativa_base_antiga(arquivo, pacote = "openxlsx")
}
\arguments{
\item{arquivo}{Local do arquivo: uma string com o endereço parcial ou completo.}

\item{pacote}{Decide se usa o openxlsx ou o readxl. O segundo dá pau com bad_alocc
às vezes mas é mais rápido.}
}
\value{
data.frame chique.
}
\description{
Lê a tratativa já com os nomes uniformes
}
