% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabelas_antigas.R
\name{le_tratativa_base_nova}
\alias{le_tratativa_base_nova}
\title{Lê várias tratativas e arquivos de processamento da base nova}
\usage{
le_tratativa_base_nova(arquivos_eventos, arquivos_processamentos = NULL)
}
\arguments{
\item{arquivos_eventos}{Local dos arquivos com as planilhas de eventos.}

\item{arquivos_processamentos}{Local dos arquivos com as planilhas de processamentos.}
}
\value{
data.frame chique.
}
\description{
Lê várias tratativas e arquivos de processamento da base nova
}
