% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabelas_antigas.R
\docType{data}
\name{nomes_colunas_tratativas}
\alias{nomes_colunas_tratativas}
\title{Vetor com os nomes das colunas das tratativas no modelo antigo.}
\format{
An object of class \code{character} of length 38.
}
\usage{
nomes_colunas_tratativas
}
\value{
Um vetor da hora.
}
\description{
Vetor com os nomes das colunas das tratativas no modelo antigo.
}
\keyword{datasets}
