% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sample_safe}
\alias{sample_safe}
\title{Retorna amostra de vetor}
\usage{
sample_safe(x = 1, size = length(x), replace = FALSE)
}
\arguments{
\item{x}{O vetor de onde tirar a amostra}

\item{size}{O tamanho da amostra}

\item{replace}{Os valores retornam para a caixa e podem ser selecionados novamente?}
}
\value{
Um vetor com a sequência
}
\description{
Versão mais sensata da função `sample`.``
}
