% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{seq_com_fim}
\alias{seq_com_fim}
\title{Gera sequência com o último termo incluso}
\usage{
seq_com_fim(from, to, by)
}
\arguments{
\item{from}{De onde}

\item{to}{Até onde}

\item{by}{De quanto em quanto}
}
\value{
Um vetor com a sequência
}
\description{
Gera sequência com o último termo incluso
}
