% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{seq_date}
\alias{seq_date}
\title{Gera sequência de dias}
\usage{
seq_date(
  from = today(tzone = "Brazil/East"),
  to = today(tzone = "Brazil/East"),
  by = 1
)
}
\arguments{
\item{from}{De onde}

\item{to}{Até onde}

\item{by}{De quantos em quantos dias}
}
\value{
Um vetor com a sequência
}
\description{
Gera sequência de dias
}
