% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set_wd_aqui}
\alias{set_wd_aqui}
\title{Define o working directory como sendo a pasta onde se encontra salvo o script.}
\usage{
set_wd_aqui()
}
\value{
Muda o wd.
}
\description{
Define o working directory como sendo a pasta onde se encontra salvo o script.
}
