% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{substitui_string}
\alias{substitui_string}
\title{Substitui um vetor de caracteres por outro caracter}
\usage{
substitui_string(x, de, para = "")
}
\arguments{
\item{x}{Um vetor de caracteres}

\item{de}{Os caracteres a serem substituídos. Pode ser um vetor.}

\item{para}{O caracter pelo qual serão trocados os valores do vetor 'de'.}
}
\description{
Substitui um vetor de caracteres por outro caracter
}
\examples{
substitui_string(x = "a < b > c", de = c("<", ">"), para = "!")

}
