% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tira_espaco_da_string}
\alias{tira_espaco_da_string}
\title{Tira espaço dos elementos de um vetor.}
\usage{
tira_espaco_da_string(x)
}
\arguments{
\item{x}{Um vetor de caracteres}
}
\value{
O vetor sem espaços.
}
\description{
Tira espaço dos elementos de um vetor.
}
\examples{
tira_espaco_do_nome("texto com espaco")
}
