% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{translate_anonymous}
\alias{translate_anonymous}
\title{Translate anonymous}
\usage{
translate_anonymous(x, translate_table)
}
\arguments{
\item{x}{To be translated.}

\item{translate_table}{Reference table, with columns Original and Anonymous}
}
\value{
Translated x
}
\description{
Given an injective table Original vs Anonymous, translates a vector to an anonymous vector.
}
