% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{troca_espaco_por_ponto}
\alias{troca_espaco_por_ponto}
\title{Troca espaço por ponto final na string}
\usage{
troca_espaco_por_ponto(x, x1)
}
\arguments{
\item{x}{Um vetor de caracteres}
}
\value{
O vetor com ponto no lugar do espaço.
}
\description{
Troca espaço por ponto final na string
}
\examples{
troca_espaco_por_ponto("texto com espaco")
}
