% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipular_strings.R
\name{troca_ponto_por_espaco}
\alias{troca_ponto_por_espaco}
\title{Troca ponto final por espaço na string}
\usage{
troca_ponto_por_espaco(x, x1)
}
\arguments{
\item{x}{Um vetor de caracteres}
}
\value{
O vetor com espaço no lugar do ponto.
}
\description{
Troca ponto final por espaço na string
}
\examples{
troca_ponto_por_espaco("texto.com.pontos")
}
