% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_com_repeticao.R
\name{try_com_repeticao}
\alias{try_com_repeticao}
\title{Tenta executar um função repetidas vezes, com tempo de espera entre elas}
\usage{
try_com_repeticao(
  expr,
  tentativas = 5,
  texto_funcao = "função x",
  tempo_espera_entre_tentativas = 0,
  valor_se_erro = NULL
)
}
\arguments{
\item{expr}{Expressão a ser executada}

\item{tentativas}{Quantidade de tentativas}

\item{tempo_espera_entre_tentativas}{O tempo de espera entre as tentativas, caso dê erro}

\item{valor_se_erro}{Valor final caso dê erro. É possível usar stop("texto") para ainda dar erro após todas as tentativas.}

\item{Texto_funcao}{O texto que será exibido caso falhe}
}
\description{
Tenta executar um função repetidas vezes, com tempo de espera entre elas
}
