% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{email_outlook}
\alias{email_outlook}
\title{Envia emails usando o Outlook}
\usage{
email_outlook(
  para = "",
  cc = "",
  bcc = "",
  assunto = "",
  texto_email = "",
  assinatura = "",
  anexos = "",
  exibir_email = TRUE,
  enviar_email = FALSE
)
}
\arguments{
\item{para}{Um vetor com emails (o que se escreve no 'para').}

\item{cc}{Um vetor com emails em cópia 'cc'.}

\item{bcc}{Um vetor com emails para cópia oculta.}

\item{assunto}{O assunto do email.}

\item{texto_email}{O texto do email (em html).}

\item{assinatura}{A assinatura do email (em html).}

\item{anexos}{Um vetor com o endereço dos anexos (o caminho pode estar relativo ao seu working directory).}

\item{exibir_email}{Se TRUE (padrão), exibe o email montado no Outlook (e voce envia apos checar).}

\item{enviar_email}{Se TRUE (o padrão é FALSE), envia o email logo após montar.}
}
\description{
Monte emails usando o R e envie com o Outlook, com anexos, múltiplos contatos, etc. O
    email pronto será exibido e bastará apertar enviar. Só funciona usando R 3.6 (e não o 4).
}
\details{
O pacote RDCOMclient é meio chato de instalar. Se ao usar essa função
ele der erro de instalação, provavelmente terá que apagar um arquivo chamado
LOCK-não-sei-o-que (conforme escrito no erro que der) e tentar usar a função de novo. É preciso
instalar o Rtools35.
}
