% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tira_ponto_da_string}
\alias{tira_ponto_da_string}
\title{Tira ponto final da string}
\usage{
tira_ponto_da_string(x, x1)
}
\arguments{
\item{x}{Um vetor de caracteres}
}
\value{
O vetor sem pontos.
}
\description{
Tira ponto final da string
}
\examples{
tira_ponto_da_string("texto.com.pontos")
}
